requirements.txt
openai-whisper
fastapi==0.115.12
uvicorn==0.34.0
langchain==0.3.25
langchain-community==0.3.23
langchain-core==0.3.58
langchain-text-splitters==0.3.8
Flask~=3.1.1
Flask-Login==0.6.3
Werkzeug~=3.1.3
ollama==0.5.1
python-dotenv==1.1.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
python-multipart==0.0.20 
aiofiles==24.1.0 
PyPDF2==3.0.0
python-docx==0.8.11
flask_sqlalchemy==3.0.2
flask_migrate==4.0.0
flask_login==0.6.3
langdetect==1.0.9
gunicorn==21.2.0